+++ kernel/include/kernel/tty.h
@@ -6,11 +6,11 @@
 int terminal_init(struct boot_table *boot_table);
 void terminal_setcolour(unsigned int fg, unsigned int bg);
 void terminal_writestring(unsigned char *data);
-void terminal_write(unsigned char *str, size_t len);
+void terminal_write(unsigned char *str, uint64_t len);
 void terminal_putchar(unsigned char c);
-void terminal_putentryat(unsigned char c, size_t column, size_t row, unsigned int fg, unsigned int bg);
+void terminal_putentryat(unsigned char c, uint64_t column, uint64_t row, unsigned int fg, unsigned int bg);
 
 void terminal_newline();
-void terminal_cursor(size_t x, size_t y);
+void terminal_cursor(uint64_t x, uint64_t y);
 
 #endif /* _TTY_H */
